+++ src/main/java/me/tomsdevsn/hetznercloud/objects/enums/ImageType.java
@@ -1,0 +1,6 @@
+package me.tomsdevsn.hetznercloud.objects.enums;
+
+public enum ImageType {
+
+    system,snapshot,backup,app
+}