+++ src/main/java/me/tomsdevsn/hetznercloud/objects/general/IPAssigneeType.java
@@ -1,6 +1,0 @@
-package me.tomsdevsn.hetznercloud.objects.general;
-
-public enum IPAssigneeType {
-
-    server
-}