+++ src/main/java/me/tomsdevsn/hetznercloud/objects/general/PlacementGroupType.java
@@ -1,5 +1,0 @@
-package me.tomsdevsn.hetznercloud.objects.general;
-
-public enum PlacementGroupType {
-    spread
-}