+++ src/main/java/me/tomsdevsn/hetznercloud/objects/request/FloatingIPRequest.java
@@ -3,7 +3,7 @@ package me.tomsdevsn.hetznercloud.objects.request;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
-import me.tomsdevsn.hetznercloud.objects.general.IPType;
+import me.tomsdevsn.hetznercloud.objects.enums.IPType;
 
 import java.util.Map;